# apps/context-gpt-backend/context-gpt.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/threads/{id}/messages": {
      "post": {
        "operationId": "ThreadsController_handleClaudeRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadsIdMessagesRequestPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/check-token": {
      "post": {
        "operationId": "AuthController_checkToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckTokenInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckTokenOutputDto"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_healthCheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthCheckOutputDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "ContextGpt API",
    "description": "API for interacting with the ContextGpt application",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "context-gpt",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "MessageDto": {
        "type": "object",
        "properties": {
          "sender": {
            "type": "string",
            "enum": [
              "User",
              "Assistant"
            ]
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "sender",
          "content"
        ]
      },
      "ThreadsIdMessagesRequestPostDto": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            }
          }
        },
        "required": [
          "messages"
        ]
      },
      "CheckTokenInputDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "CheckTokenOutputDto": {
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean"
          }
        },
        "required": [
          "isValid"
        ]
      },
      "HealthCheckOutputDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OK"
            ]
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}